+++ core/providers/cohere/types.go
@@ -293,7 +293,7 @@ type CohereEmbeddingResponse struct {
 
 // CohereEmbeddingData represents the embeddings object with different types
 type CohereEmbeddingData struct {
-	Float   [][]float32 `json:"float,omitempty"`   // Float embeddings
+	Float   [][]float64 `json:"float,omitempty"`   // Float embeddings
 	Int8    [][]int8    `json:"int8,omitempty"`    // Int8 embeddings
 	Uint8   [][]uint8   `json:"uint8,omitempty"`   // Uint8 embeddings
 	Binary  [][]int8    `json:"binary,omitempty"`  // Binary embeddings
@@ -421,8 +421,8 @@ type CohereCitation struct {
 type CohereSource struct {
 	Type       CohereSourceType `json:"type"`                  // Source type ("tool" or "document")
 	ID         *string          `json:"id,omitempty"`          // Source ID (nullable)
-	ToolOutput *json.RawMessage  `json:"tool_output,omitempty"` // Tool output (for tool sources, json.RawMessage preserves key ordering)
-	Document   *json.RawMessage  `json:"document,omitempty"`    // Document data (for document sources, json.RawMessage preserves key ordering)
+	ToolOutput *json.RawMessage `json:"tool_output,omitempty"` // Tool output (for tool sources, json.RawMessage preserves key ordering)
+	Document   *json.RawMessage `json:"document,omitempty"`    // Document data (for document sources, json.RawMessage preserves key ordering)
 }
 
 // ==================== STREAMING TYPES ====================
